requirements.txt
fastapi
uvicorn
panda
flask